Remove date field from TidalEvent
The date field was provided from the API always as UTC ("GMT"),
and the move to using Europe/London as the timezone for datetimes
meant the date could appear to be from the previous day.

Having ready access to the date of a tidal event is helpful, though,
for filtering purposes ("show me today's tides") so a new method will
return the date portion of the event datetime.