lib: re-export all of types module
The choice was between make types public, in which case there would be
re-exports of the structs people would construct or receive directly
(Station, StationId, TidePredictions) or just `pub use` everything in
lib.rs.

Really the types module is just an organisational tool for me, it
doesn't add anything for users beyond limiting the list of things that
show up in the module root documentation.